server/requirements.txt
anyio==3.6.2
APScheduler==3.6.3
asgiref==3.5.2
autopep8==2.0.0
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
convert-numbers==0.4
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==22.1
django-rest-auth==0.9.5
django-rest-knox==4.2.0
djangorestframework==3.14.0
Flask==2.2.2
geographiclib==2.0
geopy==2.3.0
geos==0.2.3
h11==0.14.0
httpcore==0.16.1
httpx==0.23.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
jdatetime==4.1.0
Jinja2==3.0.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.2
pathlib==1.0.1
Pillow==9.3.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pygame==2.1.2
PyJWT==2.6.0
python-telegram-bot==13.14
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986
ruamel-yaml==0.17.21
ryanair-py==1.0.2
six==1.16.0
sniffio==1.3.0
spatialite==0.0.3
SQLAlchemy==1.4.44
sqlparse==0.4.3
telegram==0.0.1
toml==0.10.2
tornado==6.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0